A function that prints the sum of the two diagonals of a square matrix of integers(Initialize sum)

+++ 0x07-pointers_arrays_strings/8-print_diagsums.c
@@ -13,7 +13,7 @@
 void print_diagsums(int *a, int size)
 {
 	int i, j;
-	size_t sum1 = 0, sum2 = 0;
+	int sum1 = 0, sum2 = 0;
 
 	for (i = 0; i < size; i++)
 	{
